All exercises implemented and working

+++ src/main/java/com/thoughtworks/tw101/exercises/exercise5/RectangleAverager.java
@@ -3,7 +3,7 @@ package com.thoughtworks.tw101.exercises.exercise5;
 public class RectangleAverager {
 
     public float averageArea(Rectangle[] rectangles) {
-        float sum = 0 ;
+        int sum = 0 ;
 
         for (Rectangle rect: rectangles){
             sum += rect.area() ;
